tables fixed

+++ backend/src/controllers/book.controller.js
@@ -1,10 +1,14 @@
+import path from 'path';
+import fs from 'fs/promises';
 import { db } from "../models/index.js";
+import * as url from 'url';
 const Book = db.Book;
+const __dirname = url.fileURLToPath(new URL ('.', import.meta.url));
 
 export const createBook = async (req, res) => {
     try {
         let { title, type, edition, date, author } = req.body;
-        if (!title || !date || !author) throw new Error("Missing fields");
+        if (!title || !author || !edition || !date || !author) throw new Error("Missing fields");
 
         let book = {
             title: title,
@@ -28,49 +32,52 @@ export const createBook = async (req, res) => {
 export const findAllBook = async (req, res) => {
     try {
         let books = await Book.findAll();
-        if (books == []) {
+        if (books.length === 0) {
             return res.status(404).send({ message: "No data found" });
         }
         res.status(200).send(books);
     } catch (err) {
         res.status(500).send({
-            message: "Error getting all the users",
+            message: "Error getting all the books",
         });
     }
 };
 
 export const findOneBook = async (req, res) => {
     try {
-        let { bookId } = req.params;
-        let book = await user.findOne({ where: { id: bookId } });
+        const { bookId } = req.params;
+        const book = await Book.findOne({ where: { id: bookId } });
+
         if (!book) {
             return res.status(404).send({ message: "Book Not Found!" });
         }
+
         res.status(200).send(book);
     } catch (err) {
+        console.error(err);
         res.status(500).send({ message: "Error finding the book" });
     }
 };
 
-export const updateBook = (req, res) => {
-    const id = req.params.id;
-    Book.findByPk(id)
-        .then(book => {
-            if (book.filename) {
-                const imagePath = path.join(__dirname, '../public/images/book', book.filename);
-                fs.unlink(imagePath, (err) => {
-                    if (err) console.error('Error deleting img');
-                    else { console.log('image deleted') };
-                })
-            }
-        }).catch(error => {
-            res.status(500).send({ message: 'Error finding book' });
+export const updateBook = async (req, res) => {
+    try {
+        const bookId = req.params.bookId;
+        const book = await Book.findByPk(bookId);
+
+        if (!book) {
+            return res.status(404).send({ message: "Book Not Found!" });
+        }
+
+        if (book.filename) {
+            const imagePath = path.join(__dirname, '../public/images', book.filename);
+            await fs.unlink(imagePath);
+            console.log('Image deleted');
+        }
+
+        const [num] = await Book.update(req.body, {
+            where: { id: bookId }
         });
 
-
-    Book.update(req.body, {
-        where: { id: id }
-    }).then(num => {
         if (num == 1) {
             res.send({
                 message: "Book updated."
@@ -80,29 +87,31 @@ export const updateBook = (req, res) => {
                 message: "Book cannot be updated."
             });
         }
-    }).catch(error => {
+    } catch (error) {
+        console.error(error);
         res.status(500).send({ message: 'Error updating book' });
-    });
+    }
 };
 
-export const deleteOneBook = (req, res) => {
-    const id = req.params.id;
-    Book.findByPk(id)
-        .then(article => {
-            if (article.filename) {
-                const imagePath = path.join(__dirname, '../public/images/book', book.filename);
-                fs.unlink(imagePath, (err) => {
-                    if (err) console.error('Error deleting img');
-                    else { console.log('image deleted') };
-                })
-            }
-        }).catch(error => {
-            res.status(500).send({ message: 'Error finding book' });
+export const deleteOneBook = async (req, res) => {
+    try {
+        const bookId = req.params.bookId;
+        const book = await Book.findByPk(bookId);
+
+        if (!book) {
+            return res.status(404).send({ message: "Book Not Found!" });
+        }
+
+        if (book.filename) {
+            const imagePath = path.join(__dirname, '../public/images', book.filename);
+            await fs.unlink(imagePath);
+            console.log('Image deleted');
+        }
+
+        const num = await Book.destroy({
+            where: { id: bookId }
         });
 
-    Book.destroy({
-        where: { id: id }
-    }).then(num => {
         if (num == 1) {
             res.send({
                 message: "Book deleted."
@@ -112,7 +121,9 @@ export const deleteOneBook = (req, res) => {
                 message: "Book cannot be deleted."
             });
         }
-    }).catch(error => {
+    } catch (error) {
+        console.error(error);
         res.status(500).send({ message: 'Error deleting Book' });
-    });
-}; 
+    }
+};
+
